api/v1/auth/urls.py
from django.urls import path, include
from rest_framework_simplejwt.views import TokenObtainPairView,TokenRefreshView
from api.v1.auth.views import register, change_password



urlpatterns = [
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('register/',register),
    path('change_password/',change_password)
    path('profile/', include('api.v1.profile.urls')),
]